chore(renderer): address gemini review on PR #24
Local gemini-review findings (off-PR):

- Extract Op/RecordingCtx/makeCtx into lib/canvas-recorder.ts; both
  box-drawing.test.ts and powerline.test.ts now consume the shared
  helper, dropping ~250 lines of duplicated mock-context scaffolding.
- Single `as unknown as` cast lives once inside makeRecordingCtx;
  test files import a value already typed as CanvasRenderingContext2D
  and pass it straight to draw functions, no per-test cast jump.
- Remove the closure allocated by withMirror() on every call to
  fillStadium/innerStrokeStadium/fillBisector — replaced with an
  inline `if (mirror) { save; ... }` guard around a no-allocation
  mirrorAroundCellCenter helper that just does the translate/scale.

Gemini-bot inline review on PR #24:

- demo/bin/render-test.ts:280 calculateDiffPercent now decodes both
  PNGs via fast-png (already a project dep) and compares RGBA
  pixel-by-pixel with a ±2/channel tolerance for AA jitter, instead
  of byte-comparing the compressed PNG buffers. Dimension mismatch
  reports 100% diff. Fixes the flakiness the bot flagged: PNG
  metadata or compression-level changes no longer count as visual
  regressions.

All 69 lib tests still pass; typecheck, fmt, lint clean (the two
pre-existing biome warnings in url-detection.test.ts are unchanged).
